+++ src/main/java/com/planner/cash_flow/dto/IncomeDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -14,7 +15,7 @@ public class IncomeDto {
 
     private String name;
     private int value;
-    private int valueCurrency;
+    private BigDecimal valueCurrency;
     private LocalDate date;
     private User user;
