+++ src/main/java/com/planner/cash_flow/model/Income.java
@@ -12,6 +12,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -31,7 +32,7 @@ public class Income {
 
     private int value;
 
-    private int valueCurrency;
+    private BigDecimal valueCurrency;
 
     private LocalDate date;
